Harden root buildscript for cross-project references
These changes include hiding the reference of the MCP extension in the
`:mcp` project behind providers, and changing all local (within the
buildscript) references to use the root project's extension properties
via the `ext` block and `rootProject.`.

This makes review easier when these buildscripts are split into different
files later on, so the files are only copy-pastes of the relevant blocks
in the original buildscript.