Split off project configuration from root buildscript
This change greatly reduces the size of the root buildscript, and eases
viewing of the configuration for each project, as they are visible in
each file.

Currently, these project buildscripts are direct copy-pastes from their
old section in the root buildscript. Follow-up commits will change and
modify the buildscripts. This keeps the history of the changes intact,
for ease of review.